# Fix certain binary expression queries by optimizing through push down