--- a/pkg/logql/downstream_test.go
+++ b/pkg/logql/downstream_test.go
@@ -132,14 +132,19 @@ func TestRangeMappingEquivalence(t *testing.T) {
 
 		// sum
 		{`sum(bytes_over_time({a=~".+"}[2s]))`, time.Second},
+		{`sum(bytes_over_time({a=~".+"} | logfmt | line > 5 [2s]))`, time.Second},
 		{`sum(count_over_time({a=~".+"}[2s]))`, time.Second},
+		{`sum(count_over_time({a=~".+"} | logfmt | line > 5 [2s]))`, time.Second},
 		{`sum(sum_over_time({a=~".+"} | unwrap b [2s]))`, time.Second},
+		{`sum(sum_over_time({a=~".+"} | logfmt | unwrap line [2s]))`, time.Second},
 		{`sum(max_over_time({a=~".+"} | unwrap b [2s]))`, time.Second},
 		{`sum(max_over_time({a=~".+"} | unwrap b [2s]) by (a))`, time.Second},
 		{`sum(max_over_time({a=~".+"} | logfmt | unwrap line [2s]) by (a))`, time.Second},
+		{`sum(max_over_time({a=~".+"} | logfmt | unwrap line [2s]))`, time.Second},
 		{`sum(min_over_time({a=~".+"} | unwrap b [2s]))`, time.Second},
 		{`sum(min_over_time({a=~".+"} | unwrap b [2s]) by (a))`, time.Second},
 		{`sum(min_over_time({a=~".+"} | logfmt | unwrap line [2s]) by (a))`, time.Second},
+		{`sum(min_over_time({a=~".+"} | logfmt | unwrap line [2s]))`, time.Second},
 		{`sum(rate({a=~".+"}[2s]))`, time.Second},
 		{`sum(bytes_rate({a=~".+"}[2s]))`, time.Second},
 
@@ -215,9 +220,11 @@ func TestRangeMappingEquivalence(t *testing.T) {
 		{`min(max_over_time({a=~".+"} | unwrap b [2s]))`, time.Second},
 		{`min(max_over_time({a=~".+"} | unwrap b [2s]) by (a))`, time.Second},
 		{`min(max_over_time({a=~".+"} | logfmt | unwrap line [2s]) by (a))`, time.Second},
+		{`min(max_over_time({a=~".+"} | logfmt | unwrap line [2s]))`, time.Second},
 		{`min(min_over_time({a=~".+"} | unwrap b [2s]))`, time.Second},
 		{`min(min_over_time({a=~".+"} | unwrap b [2s]) by (a))`, time.Second},
 		{`min(min_over_time({a=~".+"} | logfmt | unwrap line [2s]) by (a))`, time.Second},
+		{`min(min_over_time({a=~".+"} | logfmt | unwrap line [2s]))`, time.Second},
 		{`min(rate({a=~".+"}[2s]))`, time.Second},
 		{`min(bytes_rate({a=~".+"}[2s]))`, time.Second},
 
@@ -237,6 +244,7 @@ func TestRangeMappingEquivalence(t *testing.T) {
 		// Binary operations
 		{`bytes_over_time({a=~".+"}[3s]) + count_over_time({a=~".+"}[5s])`, time.Second},
 		{`sum(count_over_time({a=~".+"}[3s]) * count(sum_over_time({a=~".+"} | unwrap b [5s])))`, time.Second},
+		{`sum(count_over_time({a=~".+"} | logfmt | b > 2 [3s])) / sum(count_over_time({a=~".+"} [3s]))`, time.Second},
 
 		// Multi vector aggregator layer queries
 		{`sum(max(bytes_over_time({a=~".+"}[3s])))`, time.Second},